:technologist:(gh): Try to improve the Atlas DX
I'm not really happy using flake but it seems to be a better alternative
to mise for this project. However, using Nix is not without pain; I
removed all tool I use to make my commit (lazygit + czg) because `czg`
is a NodeJS using PNPM workspaces which is not currently handle by Nix
tools.

Signed-off-by: Alexandre Nicolaie <xunleii@users.noreply.github.com>