:memo:(project:amiya.akn): Update architecture for infisical
Update the architecture diagram and documentation to reflect the adoption
of Infisical for secret management.

This change introduces Infisical, PostgreSQL, and Redis to the diagram,
removes the deprecated SmallStep CA, and defines new visual styles for
datastores and management connections. The README and DRP are also
updated to align with this new architecture.

Signed-off-by: Alexandre Nicolaie <xunleii@users.noreply.github.com>